requirements.txt
asgiref==3.8.1
astroid==3.2.3
backports.csv==1.0.7
chardet==5.2.0
click==8.1.7
colorama==0.4.6
configparser==7.0.0
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
Django==5.0.7
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-import-export==4.1.1
docx-mailmerge==0.5.0
et-xmlfile==1.1.0
future==1.0.0
gunicorn==22.0.0
isort==5.13.2
jdcal==1.4.1
Jinja2==3.1.4
lazy-object-proxy==1.10.0
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==2.0.0
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
psycopg2-binary==2.9.9
pylint==3.2.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tomlkit==0.13.0
tzdata==2024.1
unicodecsv==0.14.1
wrapt==1.16.0
xlrd==2.0.1
xlwt==1.3.0
